made explore pallet its own component
