+++ client/src/components/Discover/index.js
@@ -1,0 +1,7 @@
+// Import 
+import Explore from "./Explore";
+
+// Export 
+export {
+  Explore,
+};
